+++ src/commands/appdistribution-testers-remove.ts
@@ -5,6 +5,7 @@ import { FirebaseError } from "../error";
 import { needProjectNumber } from "../projectUtils";
 import { AppDistributionClient } from "../appdistribution/client";
 import { getEmails } from "../appdistribution/options-parser-util";
+import {logger} from '../logger';
 
 module.exports = new Command("appdistribution:testers:remove [emails...]")
   .description("Remove testers")
@@ -16,18 +17,16 @@ module.exports = new Command("appdistribution:testers:remove [emails...]")
     const emailsArr = getEmails(emails, options.file);
     let deleteResponse;
     try {
-      console.log(`Deleting ${emailsArr.length} testers from project ${projectNumber}`);
+      utils.logBullet(`Deleting ${emailsArr.length} testers from project`);
       deleteResponse = await appDistroClient.removeTesters(projectNumber, emailsArr);
     } catch (err) {
-      throw new FirebaseError(`Failed to remove testers ${err}` + { exit: 1 });
+      throw new FirebaseError(`Failed to remove testers ${err}`);
     }
-    if (deleteResponse.emails != null) {
-      if (options.debug) {
-        utils.logSuccess(`Testers: ${deleteResponse.emails}, have been successfully deleted`);
-      } else {
-        utils.logSuccess(`${deleteResponse.emails.length} testers have successfully been deleted`);
-      }
-    } else {
-      utils.logSuccess(`Testers specified did not exist`);
+
+    if(!deleteResponse.emails){
+      utils.logSuccess(`Testers did not exist`);
+      return;
     }
+    logger.debug(`Testers: ${deleteResponse.emails}, have been successfully deleted`);
+    utils.logSuccess(`${deleteResponse.emails.length} testers have successfully been deleted`);
   });